requirements.txt
aiocron==2.1
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.6
aiosignal==1.4.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cronsim==2.6
exceptiongroup==1.3.0
fastapi==0.95.2
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
propcache==0.3.2
pyaes==1.6.1
pyasn1==0.6.1
pydantic==1.10.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==22.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
Telethon==1.41.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.22.0
yarl==1.20.1
